+++ src/main/java/io/nats/stan/package-info.java
@@ -1,8 +1,0 @@
-/**
- * io.nats.stan is a Java client library for the STAN streaming data system.
- * 
- * <p>This is the description?
- * 
- * @since 1.0
- */
-package io.nats.stan;