+++ src/main/java/io/nats/streaming/package-info.java
@@ -1,0 +1,14 @@
+/*
+ *  Copyright (c) 2015-2016 Apcera Inc. All rights reserved. This program and the accompanying
+ *  materials are made available under the terms of the MIT License (MIT) which accompanies this
+ *  distribution, and is available at http://opensource.org/licenses/MIT
+ */
+
+/**
+ * io.nats.streaming is a Java client library for the STAN streaming data system.
+ * 
+ * <p>This is the description?
+ * 
+ * @since 1.0
+ */
+package io.nats.streaming;